feat: Add FreeBSD support, refine cross-platform builds, and enhance audio processing
This commit introduces comprehensive support for FreeBSD (x86_64 and aarch64) and significantly refines the existing cross-platform build processes. It also includes general audio processing enhancements and repository improvements.

- **Build System & CI Enhancements:**
    - Added GitHub Actions workflows (`build-freebsd-x64`, `build-freebsd-arm64`) for building native `miniaudio` libraries for FreeBSD x86_64 and aarch64.
    - Corrected Linux cross-compilation setup within the CI pipeline:
        - Explicitly added `qemu-user-static` and relevant ARM/ARM64 GCC toolchains (`gcc-arm-linux-gnueabihf`, `gcc-aarch64-linux-gnu`) to ensure correct non-x64 binaries were produced. Previously, an issue caused x64 binaries to be generated for ARM/ARM64 targets due to missing cross-compilation tools.
        - Downgraded CI runner OS from `ubuntu-latest` to `ubuntu-22.04` to enhance compatibility with a broader range of older systems.
    - Updated CMake configurations for Android and Linux to ensure proper targeting and toolchain usage.
    - Refined platform detection and library loading logic in `Native.cs` to include FreeBSD.
    - Updated dependency analysis for Linux builds to use `readelf -d` for more robust symbol inspection.

- **Audio Engine Improvements:**
    - Refactored `AudioEngine.ConvertAndCopyToOutput` method for improved performance, robustness, and accuracy.
    - Implemented specific, optimized conversion paths for `byte`, `short`, and `int` sample formats using `System.Runtime.CompilerServices.Unsafe.As` and `MethodImplOptions.AggressiveInlining`.
    - Enhanced error handling and sample clamping to correctly manage `NaN` and `Infinity` values, preventing unexpected audio artifacts or crashes.
    - Added parameter validation to prevent common misuse.

- **Native Libraries:**
    - Rebuilt and updated `miniaudio` native libraries for all supported platforms (Windows, Linux, macOS, iOS, and the newly added FreeBSD) to incorporate the latest changes and ensure compatibility.

- **Repository Maintenance:**
    - Replaced GitHub issue templates with issue forms to streamline the issue submission experience and improve data collection.
