--- .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,82 +1,0 @@
----
-name: "\U0001F41B Bug Report"
-about: Create a report to help us improve SoundFlow
-title: "[BUG] Brief description of the bug"
-labels: bug, documentation
-assignees: ''
-
----
-
-**Thank you for reporting a bug! To help us understand and fix the issue, please provide the following information:**
-
-**Requirements:**
-*   [ ] I have searched the existing issues to ensure this bug has not already been reported.
-*   [ ] I have provided a clear and concise description of the bug.
-*   [ ] I have provided detailed steps to reproduce the bug.
-*   [ ] I have included a minimal reproducible example (MRE) if applicable.
-*   [ ] I have provided error messages and stack traces if applicable.
-*   [ ] I have completed the environment information.
-
-**1. Describe the Bug**
-
-A clear and concise description of what the bug is.
-
-**2. Steps to Reproduce**
-
-Please provide detailed steps to reproduce the behavior:
-1. Go to '...'
-2. Call method '....' with parameters '....'
-3. Configure component '....' with settings '....'
-4. See error / Observe unexpected behavior
-
-**3. Expected Behavior**
-
-A clear and concise description of what you expected to happen.
-
-**4. Current Behavior**
-
-A clear and concise description of what actually happens.
-
-**5. Minimal Reproducible Example (MRE)**
-
-Please provide a *minimal* code snippet that demonstrates the bug. This is crucial for us to quickly identify and fix the issue.
-If the code is too long, please link to a Gist or a small repository.
-
-```csharp
-// Your MRE code here
-```
-
-**6. Error Messages and Stack Trace (if applicable)**
-
-If an exception is thrown, please paste the full error message and stack trace below.
-
-```
-(Paste full error message and stack trace here)
-```
-
-**7. Environment (please complete the following information):**
-*   **SoundFlow Version:** [e.g., 1.0.0, or commit SHA if building from source]
-*   **.NET Version:** [e.g., .NET 8.0, .NET 9.0]
-*   **Operating System:** [e.g., Windows 11, macOS Sonoma 14.2, Ubuntu 22.04]
-*   **Architecture:** [e.g., x64, ARM64]
-*   **Audio Backend Used (if known):** [e.g., WASAPI, CoreAudio, etc.]
-*   **Specific Audio Hardware (if relevant):** [e.g., Realtek HD Audio, Focusrite Scarlett 2i2]
-
-**8. Affected Components/Modules (if known)**
-
-List any specific SoundFlow components, modifiers, providers, or extensions you believe are related to this bug.
-*   [e.g., `SoundPlayer`, `WebRtcApmModifier`, `MiniAudioDecoder`, `NetworkDataProvider`]
-
-**9. Impact**
-
-Describe the impact of this bug (e.g., crash, incorrect audio output, performance degradation, minor inconvenience).
-
-**10. Possible Workaround (if known)**
-
-If you've found a way to work around the issue, please describe it.
-
-**11. Additional Context**
-
-Add any other context about the problem here.
-*   Were there any recent changes to your environment or the library version?
-*   Is this bug consistently reproducible, or does it happen intermittently?
