--- .github/ISSUE_TEMPLATE/documentation_issue.md
@@ -1,57 +1,0 @@
----
-name: "\U0001F4D6 Documentation Issue"
-about: Report an issue or suggest an improvement for the documentation
-title: "[DOCS] Brief description of the documentation issue"
-labels: documentation
-assignees: ''
-
----
-
-**Thank you for helping us improve the SoundFlow documentation!**
-
-**Requirements:**
-*   [ ] I have searched the existing issues to ensure this documentation issue has not already been reported.
-*   [ ] I have clearly identified the affected documentation.
-*   [ ] I have described the problem and proposed a clear improvement.
-
-**1. Affected Page/Section**
-
-Please provide the URL or a clear description of the documentation page or section that has an issue.
-*   URL (if applicable):
-*   Section/Topic: [e.g., "SoundPlayer Looping", "WebRtcApmModifier Configuration", "API reference for Mixer.AddComponent"]
-
-**2. Describe the Issue**
-
-A clear and concise description of what is wrong with the current documentation.
-*   Is something missing?
-*   Is something unclear or confusing?
-*   Is something incorrect or outdated?
-*   Are there typos or grammatical errors?
-
-**3. Suggested Improvement**
-
-Please describe how the documentation could be improved.
-*   What information should be added, changed, or removed?
-*   Provide a corrected version if possible.
-
-**4. Why is this Improvement Important?**
-
-Explain how this change would make the documentation better or easier to understand for users.
-
-**5. Code Examples (if relevant)**
-
-If the documentation issue relates to code examples, please provide the problematic example and/or a corrected/improved version.
-
-**Current Example (if applicable):**
-```csharp
-// Problematic code example
-```
-
-**Suggested Example (if applicable):**
-```csharp
-// Improved code example
-```
-
-**6. Additional Context**
-
-Add any other relevant information, screenshots, or links.
