+++ .github/ISSUE_TEMPLATE/documentation_issue.yaml
@@ -1,0 +1,101 @@
+name: "📚 Documentation Issue"
+description: Report an issue or suggest an improvement for the documentation
+title: "[DOCS] Brief description of the documentation issue"
+labels: ["documentation"]
+assignees: []
+
+body:
+  - type: markdown
+    attributes:
+      value: |
+        **Thank you for helping us improve the SoundFlow documentation!**
+
+  - type: checkboxes
+    id: requirements
+    attributes:
+      label: Requirements
+      options:
+        - label: I have searched the existing issues to ensure this documentation issue has not already been reported.
+          required: true
+        - label: I have clearly identified the affected documentation.
+          required: true
+        - label: I have described the problem and proposed a clear improvement.
+          required: true
+
+  - type: textarea
+    id: affected-page
+    attributes:
+      label: 1. Affected Page/Section
+      description: |
+        Please provide the URL or a clear description of the documentation page or section that has an issue.
+        *   URL (if applicable):
+        *   Section/Topic: [e.g., "SoundPlayer Looping", "WebRtcApmModifier Configuration", "API reference for Mixer.AddComponent"]
+      placeholder: |
+        URL: https://docs.soundflow.com/some-page
+        Section: Introduction to SoundPlayer
+    validations:
+      required: true
+
+  - type: textarea
+    id: describe-issue
+    attributes:
+      label: 2. Describe the Issue
+      description: |
+        A clear and concise description of what is wrong with the current documentation.
+        *   Is something missing?
+        *   Is something unclear or confusing?
+        *   Is something incorrect or outdated?
+        *   Are there typos or grammatical errors?
+      placeholder: Describe the problem with the documentation here.
+    validations:
+      required: true
+
+  - type: textarea
+    id: suggested-improvement
+    attributes:
+      label: 3. Suggested Improvement
+      description: |
+        Please describe how the documentation could be improved.
+        *   What information should be added, changed, or removed?
+        *   Provide a corrected version if possible.
+      placeholder: How can the documentation be made better?
+    validations:
+      required: true
+
+  - type: textarea
+    id: importance
+    attributes:
+      label: 4. Why is this Improvement Important?
+      description: Explain how this change would make the documentation better or easier to understand for users.
+      placeholder: Why is this change necessary?
+    validations:
+      required: true
+
+  - type: textarea
+    id: code-examples
+    attributes:
+      label: 5. Code Examples (if relevant)
+      description: |
+        If the documentation issue relates to code examples, please provide the problematic example and/or a corrected/improved version.
+      placeholder: |
+        Current Example (if applicable):
+        ```csharp
+        // Problematic code example
+        ```
+
+        Suggested Example (if applicable):
+        ```csharp
+        // Improved code example
+        ```
+      render: shell
+    validations:
+      required: false
+
+  - type: textarea
+    id: additional-context
+    attributes:
+      label: 6. Additional Context
+      description: Add any other relevant information, screenshots, or links.
+      placeholder: Any other relevant information.
+    validations:
+      required: false
