--- .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,54 +1,0 @@
----
-name: "✨ Feature Request"
-about: Suggest an idea or enhancement for SoundFlow
-title: "[FEAT] Brief description of the feature"
-labels: enhancement
-assignees: ''
-
----
-
-**Thank you for suggesting a feature! We appreciate your input.**
-
-**Requirements:**
-*   [ ] I have searched the existing issues to ensure this feature has not already been requested.
-*   [ ] I have clearly described the problem and the proposed solution.
-*   [ ] I have considered alternative solutions.
-
-**1. Is your feature request related to a problem? Please describe.**
-
-A clear and concise description of what the problem is.
-
-*Example: I'm always frustrated when [...] because [...] and I would like to be able to [...]*
-
-**2. Describe the Solution You'd Like**
-
-A clear and concise description of what you want to happen.
-*   What is the desired behavior?
-*   How would this feature be used? (e.g., API examples, new component, modification to existing one)
-
-**3. Describe Alternatives You've Considered**
-
-A clear and concise description of any alternative solutions or features you've considered.
-*   Why were they not suitable?
-
-**4. Proposed API (if applicable)**
-
-If this involves new or changed API, please provide a sketch of how it might look.
-```csharp
-// public class NewComponent : SoundComponent { ... }
-// public void ExistingComponent.NewMethod(ParameterType param) { ... }
-```
-
-**5. Benefits**
-
-What are the benefits of implementing this feature?
-*   Who would benefit from this?
-*   How would it improve SoundFlow?
-
-**6. Potential Drawbacks/Challenges**
-
-Are there any potential downsides or implementation challenges you foresee?
-
-**7. Additional Context**
-
-Add any other context, examples, mockups, or links to relevant resources or similar features in other libraries.
