--- .github/ISSUE_TEMPLATE/question.md
@@ -1,47 +1,0 @@
----
-name: "\U0001F914 Question"
-about: Ask a question about SoundFlow
-title: "[QSTN] Your question in a few words"
-labels: question
-assignees: ''
-
----
-
-**Before asking, please ensure your question isn't answered in the [documentation](https://lsxprime.github.io/soundflow-docs/) or existing issues.**
-
-**Requirements:**
-*   [ ] I have searched the documentation and existing issues for an answer.
-*   [ ] My question is specific and clearly formulated.
-*   [ ] This is not a bug report or a feature request.
-
-**1. What are you trying to achieve?**
-
-A clear and concise description of your goal.
-
-**2. What have you tried so far?**
-
-Please describe the steps you've taken, what you've looked at in the documentation, or any code you've experimented with.
-
-**3. What are you unsure about / What is your specific question?**
-
-Be precise.
-
-**4. Relevant Code Snippet (if applicable)**
-
-If your question relates to a specific piece of code, please include a minimal, runnable snippet.
-```csharp
-// Your code here
-```
-
-**5. Environment (if relevant to your question):**
-*   **SoundFlow Version:** [e.g., 1.0.0]
-*   **.NET Version:** [e.g., .NET 8.0]
-*   **Operating System:** [e.g., Windows 11]
-
-**6. Additional Context**
-
-Any other information that might help us understand and answer your question.
-
-**Important:**
-*   This issue tracker is primarily for bugs and feature requests. For general "how-to" questions, please use the [Discussions tab](https://github.com/LSXPrime/SoundFlow/discussions) if available. However, if your question might reveal a documentation gap or an unintuitive aspect of the library, this is the right place.
-*   **This is NOT for reporting bugs or requesting features.** Please use the appropriate templates for those.
