--- .github/ISSUE_TEMPLATE/security_vulnerability_report.md
@@ -1,73 +1,0 @@
----
-name: "\U0001F6E1️ Security Vulnerability Report"
-about: Report a security vulnerability in SoundFlow (Please Read Instructions!)
-title: "[SEC] Brief, non-specific description of the potential vulnerability area"
-labels: ''
-assignees: ''
-
----
-
-**⚠️ IMPORTANT: Reporting Security Vulnerabilities ⚠️**
-
-Thank you for taking the time to responsibly disclose a potential security vulnerability.
-
-**To protect our users, please DO NOT report security vulnerabilities through public GitHub issues.**
-
-Instead, please send a detailed email to: **[My Personal Email](mailto:lsxprime@gmail.com)**
-
-If you are unable to report privately or believe the issue is not sensitive enough to require private disclosure, you may proceed with this template, but be mindful of the information you share publicly.
-
-**If you are reporting privately via email, please include all the information requested below.**
-
----
-
-**(Only fill out below if you are certain public disclosure is appropriate or have been asked to do so by a maintainer)**
-
-**1. Vulnerability Type**
-
-A brief classification of the vulnerability.
-*   [e.g., Remote Code Execution (RCE), Denial of Service (DoS), Information Disclosure, Unsafe Native Code Invocation, Path Traversal in File Operations]
-
-**2. Affected Component(s)/Module(s)**
-
-Which part(s) of SoundFlow are affected?
-*   [e.g., `NetworkDataProvider`, `MiniAudioDecoder` (if dealing with malformed files), a specific `SoundModifier` that processes external input]
-
-**3. Description of Vulnerability**
-
-A clear and concise description of the vulnerability.
-*   Explain the weakness and how it can be exploited.
-
-**4. Steps to Reproduce (Proof of Concept - PoC)**
-
-Provide detailed steps or a code snippet (PoC) to demonstrate the vulnerability.
-**Be extremely careful with the PoC you share publicly.** If it's a simple exploit, consider describing it abstractly or waiting for private communication.
-1.
-2.
-3.
-
-```csharp
-// PoC code (if safe to share publicly)
-```
-
-**5. Impact**
-
-What is the potential impact if this vulnerability is exploited?
-*   [e.g., Application crash, arbitrary code execution, sensitive data leakage]
-
-**6. Suggested Mitigation (if known)**
-
-If you have an idea on how to fix or mitigate this vulnerability, please share it.
-
-**7. Environment (if specific):**
-*   **SoundFlow Version:** [e.g., 1.0.0, or commit SHA]
-*   **.NET Version:** [e.g., .NET 8.0]
-*   **Operating System:** [e.g., Windows 11]
-
-**8. Acknowledgement of Responsible Disclosure Policy**
-*   [ ] I understand the importance of responsible disclosure and have considered private reporting first (or this issue is deemed non-sensitive for public view).
-
-**Checklist (if reporting publicly):**
-*   [ ] I have considered the sensitivity of the information I am sharing.
-*   [ ] I have described the vulnerability and its potential impact.
-*   [ ] I have provided steps to reproduce if safe to do so publicly.
