+++ .github/ISSUE_TEMPLATE/security_vulnerability_report.yaml
@@ -1,0 +1,154 @@
+name: "🛡️ Security Vulnerability Report"
+description: Report a security vulnerability in SoundFlow (Please Read Instructions!)
+title: "[SEC] Brief, non-specific description of the potential vulnerability area"
+labels: []
+assignees: []
+
+body:
+  - type: markdown
+    attributes:
+      value: |
+        **⚠️ IMPORTANT: Reporting Security Vulnerabilities ⚠️**
+
+        Thank you for taking the time to responsibly disclose a potential security vulnerability.
+
+        **To protect our users, please DO NOT report security vulnerabilities through public GitHub issues.**
+
+        Instead, please send a detailed email to: **[My Personal Email](mailto:lsxprime@gmail.com)**
+
+        If you are unable to report privately or believe the issue is not sensitive enough to require private disclosure, you may proceed with this template, but be mindful of the information you share publicly.
+
+        **If you are reporting privately via email, please include all the information requested below.**
+
+        ---
+
+        **(Only fill out below if you are certain public disclosure is appropriate or have been asked to do so by a maintainer)**
+
+  - type: input
+    id: vulnerability-type
+    attributes:
+      label: 1. Vulnerability Type
+      description: |
+        A brief classification of the vulnerability.
+        e.g., Remote Code Execution (RCE), Denial of Service (DoS), Information Disclosure, Unsafe Native Code Invocation, Path Traversal in File Operations
+      placeholder: e.g., Remote Code Execution (RCE)
+    validations:
+      required: false
+
+  - type: textarea
+    id: affected-components
+    attributes:
+      label: 2. Affected Component(s)/Module(s)
+      description: |
+        Which part(s) of SoundFlow are affected?
+        e.g., `NetworkDataProvider`, `MiniAudioDecoder` (if dealing with malformed files), a specific `SoundModifier` that processes external input
+      placeholder: e.g., `NetworkDataProvider`
+    validations:
+      required: false
+
+  - type: textarea
+    id: description-of-vulnerability
+    attributes:
+      label: 3. Description of Vulnerability
+      description: |
+        A clear and concise description of the vulnerability.
+        Explain the weakness and how it can be exploited.
+      placeholder: Describe the vulnerability and how it can be exploited.
+    validations:
+      required: false
+
+  - type: textarea
+    id: steps-to-reproduce
+    attributes:
+      label: 4. Steps to Reproduce (Proof of Concept - PoC)
+      description: |
+        Provide detailed steps or a code snippet (PoC) to demonstrate the vulnerability.
+        **Be extremely careful with the PoC you share publicly.** If it's a simple exploit, consider describing it abstractly or waiting for private communication.
+      placeholder: |
+        1. ...
+        2. ...
+        3. ...
+        ```csharp
+        // PoC code (if safe to share publicly)
+        ```
+      render: shell
+    validations:
+      required: false
+
+  - type: textarea
+    id: impact
+    attributes:
+      label: 5. Impact
+      description: |
+        What is the potential impact if this vulnerability is exploited?
+        e.g., Application crash, arbitrary code execution, sensitive data leakage
+      placeholder: Describe the potential impact.
+    validations:
+      required: false
+
+  - type: textarea
+    id: suggested-mitigation
+    attributes:
+      label: 6. Suggested Mitigation (if known)
+      description: If you have an idea on how to fix or mitigate this vulnerability, please share it.
+      placeholder: How might this vulnerability be fixed or mitigated?
+    validations:
+      required: false
+
+  - type: markdown
+    attributes:
+      value: |
+        ### 7. Environment (if specific):
+
+  - type: input
+    id: sf-version
+    attributes:
+      label: SoundFlow Version
+      description: e.g., 1.0.0, or commit SHA
+    validations:
+      required: false
+
+  - type: input
+    id: net-version
+    attributes:
+      label: .NET Version
+      description: e.g., .NET 8.0
+    validations:
+      required: false
+
+  - type: input
+    id: os
+    attributes:
+      label: Operating System
+      description: e.g., Windows 11
+    validations:
+      required: false
+
+  - type: checkboxes
+    id: responsible-disclosure-ack
+    attributes:
+      label: 8. Acknowledgement of Responsible Disclosure Policy
+      options:
+        - label: I understand the importance of responsible disclosure and have considered private reporting first (or this issue is deemed non-sensitive for public view).
+          required: true
+    validations:
+      required: true
+
+  - type: markdown
+    attributes:
+      value: |
+        **Checklist (if reporting publicly):**
+
+  - type: checkboxes
+    id: public-reporting-checklist
+    attributes:
+      label: Public Reporting Checklist (Optional, only if proceeding publicly)
+      options:
+        - label: I have considered the sensitivity of the information I am sharing.
+          required: false
+        - label: I have described the vulnerability and its potential impact.
+          required: false
+        - label: I have provided steps to reproduce if safe to do so publicly.
+          required: false
+    validations:
+      required: false
